+++ apps/desktop/src/routes/_protected/database/$id/sql/-components/chat/index.ts
@@ -3,7 +3,7 @@ import type { AppUIMessage } from '@conar/api/ai/tools/helpers'
 import type { InferToolInput, InferToolOutput } from 'ai'
 import type { chatsMessages, connections } from '~/drizzle'
 import { Chat } from '@ai-sdk/react'
-import { SQL_FILTERS_LIST } from '@conar/shared/filters/sql'
+import { SQL_FILTERS_LIST } from '@conar/shared/filters'
 import { eventIteratorToStream } from '@orpc/client'
 import { encode } from '@toon-format/toon'
 import { lastAssistantMessageIsCompleteWithToolCalls } from 'ai'
