+++ apps/desktop/src/routes/_protected/database/$id/table/-components/filters/filters-provider.tsx
@@ -1,0 +1,19 @@
+import type { Filter, FilterGroup } from '@conar/shared/filters'
+import type { Column } from './context'
+import { FilterInternalContext } from './context'
+
+export function FiltersProvider({
+  children,
+  columns,
+  filtersGrouped,
+}: {
+  children: React.ReactNode
+  columns: Column[]
+  filtersGrouped: { group: FilterGroup, filters: Filter[] }[]
+}) {
+  return (
+    <FilterInternalContext value={{ columns, filtersGrouped }}>
+      {children}
+    </FilterInternalContext>
+  )
+}
