+++ packages/table/tsconfig.json
@@ -1,11 +1,6 @@
 {
   "extends": "@conar/configs/ts/react.json",
   "compilerOptions": {
-    "paths": {
-      "@conar/table/*": ["./src/*"]
-    },
     "noEmit": true
-  },
-  "include": ["."],
-  "exclude": ["node_modules"]
+  }
 }
